+++ src/basic/App.tsx
@@ -19,6 +19,8 @@ interface Notification {
 
 import { useCart } from './hooks/useCart';
 
+import { useCoupons } from './hooks/useCoupons';
+
 const App = () => {
   const [notifications, setNotifications] = useState<Notification[]>([]);
   const addNotification = useCallback(
@@ -35,9 +37,7 @@ const App = () => {
 
   const { products, addProduct, updateProduct, deleteProduct } = useProducts();
   const { cart, setCart, addToCart, removeFromCart, updateQuantity, getRemainingStock } = useCart(addNotification);
-  const [coupons, setCoupons] = useLocalStorage<Coupon[]>('coupons', initialCoupons);
-
-  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
+  const { coupons, selectedCoupon, setSelectedCoupon, addCoupon, deleteCoupon, applyCoupon } = useCoupons(cart, addNotification);
   const [isAdmin, setIsAdmin] = useState(false);
   const [showCouponForm, setShowCouponForm] = useState(false);
   const [activeTab, setActiveTab] = useState<'products' | 'coupons'>(
@@ -77,26 +77,6 @@ const App = () => {
     discountValue: 0,
   });
 
-  
-
-  const applyCoupon = useCallback(
-    (coupon: Coupon) => {
-      const currentTotal = calculateCartTotal(cart, selectedCoupon).totalAfterDiscount;
-
-      if (currentTotal < 10000 && coupon.discountType === 'percentage') {
-        addNotification(
-          'percentage 쿠폰은 10,000원 이상 구매 시 사용 가능합니다.',
-          'error'
-        );
-        return;
-      }
-
-      setSelectedCoupon(coupon);
-      addNotification('쿠폰이 적용되었습니다.', 'success');
-    },
-    [addNotification, calculateCartTotal]
-  );
-
   const completeOrder = useCallback(() => {
     const orderNumber = `ORD-${Date.now()}`;
     addNotification(
@@ -106,32 +86,6 @@ const App = () => {
     setCart([]);
     setSelectedCoupon(null);
   }, [addNotification]);
-
-  
-
-  const addCoupon = useCallback(
-    (newCoupon: Coupon) => {
-      const existingCoupon = coupons.find((c) => c.code === newCoupon.code);
-      if (existingCoupon) {
-        addNotification('이미 존재하는 쿠폰 코드입니다.', 'error');
-        return;
-      }
-      setCoupons((prev) => [...prev, newCoupon]);
-      addNotification('쿠폰이 추가되었습니다.', 'success');
-    },
-    [coupons, addNotification]
-  );
-
-  const deleteCoupon = useCallback(
-    (couponCode: string) => {
-      setCoupons((prev) => prev.filter((c) => c.code !== couponCode));
-      if (selectedCoupon?.code === couponCode) {
-        setSelectedCoupon(null);
-      }
-      addNotification('쿠폰이 삭제되었습니다.', 'success');
-    },
-    [selectedCoupon, addNotification]
-  );
 
   const handleProductSubmit = (e: React.FormEvent) => {
     e.preventDefault();
